pypi: Update build triggers for PyPI builds on Azure
The PyPI pipeline no longer triggers for pull requests
unless the PR touches PyPI-related configuration. These
builds were typically unnecessary since they almost
always matched the results of the regular pipeline builds.
Removing them reduces the load on our limited queue slots
and improves turnaround times.

It's possible (but unlikely) that a PR would break the
PyPI build but not the regular build. To cover that case,
the PyPI pipeline now runs on all updates to the dev and
release branches. This way we'll at least see the failure
when the PR is merged.

(Internal change: 2296034)