Update hd for the portal light.
Create a token for the "portals" relationship, and make sure the
relationship gets added to the light data source starting with the
HdLegacyPrimSceneIndex (where the other components of the light
data source get added). If the relationship doesn't exist (it's
only on dome lights), the light data source won't end up with a
"portals" component.

Later in the scene index network, the HdPrman portal light scene index
will depend on the "portals" data being around for domes with portals.

(Internal change: 2296666)