[usdMtlx, hdMtlx, hdSt] Process authored colorspaces by hdMtlx and Storm
This change passes the colorspace attribute on inputs affected by
colorspaces. These inputs are: color3 and color4 inputs as well as
filename inputs for image nodes returning color3 and color4 types.
Colorspaces were getting lost when the HdMaterialNetwork was
created, they are now added as a separate HdMaterialNode
parameter of the form 'colorSpace_<inputName>'. This allows
for HdMtlx to catch and apply those colorspaces to the appropriate
inputs in the re-created mtlx file used for shaderGen.

In order for the colorspace transforms to be generated in the
MaterialX shaders, the default source colorspace needs to also
be set. This is the equivalent to setting the colorspace at the
document level. This information is currently stored in the
stage metadata, which does not compose across references.
Which means that the information is lost. A schema is planned
to be added shortly, so in the meantime we are setting the
default source colorspace in Storm to be lin_rec709.

This also sets the default colorspace for MaterialX material textures to be 'raw' instead of 'auto'.
Note that the baselines that are getting lighter are due to the
change to 'raw' and the ones without changes are because of
colorspace specifications to the mtlx file.

Fixes #1878
Fixes #1523

(Internal change: 2297176)