cmake: Fix list deduplication in _pxr_target_link_libraries
We had previously been using CMake's regex MATCHES to determine
if an element was in a list. This failed for entries that were
generator expressions, like "$<INSTALL_INTERFACE:include>".
This in turn caused duplicate entries to be added to the list
of include paths which led to N^2 behavior in this code in
monolithic builds.

Prior to this fix, running just the cmake step for a
monolithic build took roughly 2 minutes on an internal test
machine. After this fix, this time dropped to 5 seconds.

Note there are other uses of MATCHES to find list elements
that need to be examined and possibly fixed.

(Internal change: 2297330)