cmake: Fix builds from or to paths containing whitespace
Prior to this change, builds where the source or target
install directory contained whitespace would fail. This
was because quotes were improperly handled when setting
up the copyHeaderForBuild custom commands. In addition,
the TO_NATIVE_PATH conversion in the command would apply
unwanted escapes to the whitespace in the path. This
would ultimately cause the file command to read and
write to invalid paths.

Note that tests in these cases still fail and will
be addressed in another fix.

(Internal change: 2297331)