Add copy on write functionality to UsdUtils_WritableLocalizationDelegate
This change makes the default behavior for USDZ package generation to use copy on write when modifying asset paths.  In the case that a path needs to be updated, an anonymous copy of the source layer will first be created and the changes will be written into it.  Users should use the  `UsdUtils_WritableLocalizationDelegate::GetLayerUsedForWriting` method to retrieve the correct layer for localization.

Note: usdzip still edits layers in place to maintain current behavior and performance.

(Internal change: 2297829)