Github #2582: Build fails - UIC not found.
Previously searching for uic was disabled because there are two different tools named uic, and we need the python oriented one.

The use of the word uic causes occasional confusion amongst end users since they may have it, yet the build configuration step says that it wasn't found. This change rewords the message slightly so as not to imply that a tool named specifically uic was expected.

Fixes #2582

(Internal change: 2298198)