usdGeom: skip bounds checking when there are no indices in the subset family at a particular time
This avoids undefined behavior in case indicesInFamily is empty when the bounds
checking would try to access values by de-referencing rbegin() and begin() of
the empty set. The subset family can still be considered valid if there are no
indices at a particular time as long as there are indices at some time.