[usdImaging] Update DataSourceMaterial for colorspace handling
Note that when the colorspace is added to a material input
that is connected to another node output UsdImaging and
MaterialX both ignore that colorspace specification.

I opted for keeping the GetValueProducingAttribute() instead
of swapping for input.GetAttr() because when colorspaces
were specified on materials or surfaceshaders the colorspace
was not attached to the returned attribute where it was with
GetValueProducingAttribute.

(Internal change: 2298676)