hdsi: Fix improper vector erase call
Passing a single iterator to std::vector::erase() removes the element at
that iterator. If that iterator is the past-the-end iterator, a crash can
result since the passed iterator must be dereferenceable. In this case that
happened when the names vector was empty, or the searched for element was
not found.

We now use the ranged version of erase(), which handles this case correctly
(as a no-op).

(Internal change: 2299505)