[hgiVulkan] Don't gather and store clear values within HgiVulkanGraphicsPipeline, but rather HgiVulkanGraphicsCmds.
Clear values are not actually used in HgiVulkanGraphicsPipeline. In Vulkan,
clear values are used when beginning a render pass, which is done in
HgiVulkanGraphicsCmds, which has the same information necessary to gather the
clear values itself. Additionally, storing them in HgiVulkanGraphicsPipeline
can result in incorrect drawing if we wish to CHANGE the clear values but
otherwise use the same pipeline state. In such a scenario, we would end up
using the same HgiVulkanGraphicsPipeline object and thus same clear values,
since the clear values are (rightfully) not part of the HgiGraphicsPipeline
hash.

(Internal change: 2299693)