[hgiVulkan] Add codepath for clearing attachments in HgiVulkanGraphicsCmds without beginning a render pass.
In HgiVulkan, we typically start a render pass when we are drawing something,
and render passes are highly intertwined with graphics pipelines. In situations
in which we don't bind a graphics pipeline or draw anything, we still need a
mechanism to clear the attachments if desired. This change adds a method
in HgiVulkanGraphicsCmds for doing so.

The aforementioned situation can happen during the skydome task (which when
using usdview is generally the first of the render tasks to run and thus
contains the clear information) when we don't actually want to draw the
skydome. It can also simply happen when there are no draw items in a draw batch
and we early-out of drawing.

(Internal change: 2299703)