[hd, usdImaging] Add HdMaterialNodeParameterSchema to hold the parameter value, and Color Space metadata
This replaces the single VtValue for parameter values with the
HdMaterialNodeParameterSchema which includes the VtValue
and an associated colorSpace. Unfortunately we still need to
add a 'colorSpace:inputName' parameter to the material
network to carry the colorspace for an input for Hydra 1.0.

There is no visual difference between the HdPrman tests
because the textures were already using the 'auto' color
space. Note that for HdPrman colorspaces authored on
the terminal node will not yet work in HdPrman because
we rely on PxrSurface to provide the terminal node
shading. This means that all the colorspace adjustments
supported in Storm are not supported in HdPrman.

(Internal change: 2300625)