Fix HdPrman light filter coordinate systems
The previous code was creating Riley coordinate systems for
light filters while populating lights.	Later, when Sync()'ing
subsequent changes, it would not modify those coordinate systems;
instead it would create new ones with the same names, and without
deleting the old ones.  It would also create redundant coordinate
sytsems in the case of light filters shared by multiple lights,
since each light would independently create all the coordinate
systems it uses.

Fix this by moving the responsibility for managing Riley coordinate
systems for light filters to HdPrmanLightFilter.

We need to create Riley coordinate systems for light filters
prior to using them in lights (so as to have a valid Riley ID
to refer to them).  However, Hydra does not provide an
innate way to enforce a dependency ordering among sprims,
so we use a similar approach as with how gprims in HdPrman
depend on material sprims:  the consuming HdPrman object
(here the light) requests that the depended-upon object sync
itself to Riley (SyncToRiley).

This is a regrettable complication, however we currently
expect this will be revisited with HdPrman2, when it gets
redesigned to directly use the scene index.

Also, update HdDatSourceLegacyPrim so that it provides an
xform schema for light filters.

(Internal change: 2301021)