Fix combiner behavior for light filters
When multiple light filters with the same name were used
by a light, the RenderMan combiner node would face a name
collision and be unable to distinguish among these.

Use the full path to distinguish the shading nodes for
light filters so that they can have unambiguous references
in a combiner context.

(Internal change: 2301074)