build_usd.py: Remove debug postfix from OpenImageIO library
The build script would fail to build USD in debug mode with
OpenImageIO enabled. This was a regression due to the version
upgrades in the script: OIIO added "_d" as the default postfix
when building in debug mode in v2.3.5, and USD's build
system currently doesn't know to look for the OIIO library
under a different name.

As a workaround, build_usd.py forces the debug postfix back
to its previous empty value. In the long term, we should
set up the build to rely on the cmake package exported by
the OIIO build instead of our bespoke FindOpenImageIO module.

Fixes #2749

(Internal change: 2301140)