pcp: Remove unused function parameters
The various composition task functions all took a
PcpPrimIndex pointer but only used that parameter
in a few cases. The same pointer is available via
the Pcp_PrimIndexer object that is also passed to
these functions, so we don't need to pass it in
as a separate argument.

(Internal change: 2302639)