Fix Empty output for files that are visible in usdview
When the item is really small, the camera transform is not translated
back enough to make sure the object is not culled by the near plane.
The calculated distance is intended to back the camera up to frame
the front face of the bbox however when we have small objects, that
distance value can be less than the near plane, which means the
object is culled and not visible in the generated images.

This change adds a check on the camera translation distance, so that
if it is smaller than the min clipping range (which is always the default
value of 1) then we increase the distance by 1 to make sure that the
camera is backed up enough such that the object is not culled by the
near plane. This is very similar check and adjustment to what is in
the usdview code that frames the camera around a selection.

Fixes #2556

(Internal change: 2303281)