pcp: Defer checks for unresolved prim path errors
Pcp now checks for reference and payload arcs that specify prims
that provide no opinions ("unresolved prim paths") at the very end
of prim indexing. This is a precursor for an upcoming change to the
composition algorithm.

These checks were previously done right when the reference/payload
arc was added to the prim index. This worked as expected because
all possible opinions contributing to the referenced prim would
be added to the prim index at that time. In particular, this
included opinions from ancestral variants for subroot prims. An
example of this is captured in the PrimWithReferencesInVariants
test case in the BasicReferences museum test.

An upcoming change will defer the addition of those ancestral
variant opinions until *after* the reference/payload arc is
added. This would cause unresolved prim path errors to be
incorrectly issued if we left the check where it was. Moving
the check now will avoid that problem, since it will again
ensure all opinions are available when the check occurs.

An alternate approach I considered was to leave the checks
in place and allow the incorrect errors to be recorded, but
to rescan those errors at the end of prim indexing to see
if they had been resolved by later additions to the index.
That would potentially be less risky since it'd be a smaller,
additive change. Ultimately, I decided that that approach was
just working around the issue, and that the changes in this
commit address the root cause.

(Internal change: 2304195)