[hd] Suppress dirty notices for self-dependencies after prim removal
When a prim is removed, the dependency forwarding scene index sends dirty notices to all prims that depend on the removed prim. However, when the dependencies are self-dependencies, sending dirty notices to the removed prim can lead to crashes downstream if consumers neglect to check for a null data source handle.

Scene indices should continue to always check that data source handles are valid before dereferencing them. Since self-dependency dirty-on-removal notices are not useful or actionable, this change increases signal-to-noise ratio of notices in the scene index stack.

(Internal change: 2304621)