[hdSt] Updated OSD codeGen mixins for v3.6.0
Updated Storm codeGen to use improved OpenSubdiv shader source
mixins when building with recent versions of OpenSubdiv.

This eliminates legacy shader resource declarations that are
incompatible with compiling GLSL to SPIR-V for Vulkan.

These changes are version guarded to maintain compatibility
with earlier versions of OpenSubdiv for those not yet working
with Vulkan.

(Internal change: 2305293)