[usd] Updated OpenUSD to OpenSubdiv v3.6.0
This is the version specified for the CY2024 VFX Reference Platform.

This is ahead of other packages which are currently aligned with
the CY2022 platform, but updating to this OpenSubdiv version is
helpful for ongoing hgiVulkan development.

The corresponding Storm source code changes are version guarded to
maintain compatibility with earlier versions of OpenSubdiv for
developers not yet using Vulkan.

(Internal change: 2305385)