[hdPrman] Avoid concurrent protoMap access
Under certain conditions, unprotected levels of the instancer proto map may be accessed simultaneously from multiple threads. This can lead to crashes when multiple threads simultaneously delete the same instance. This change introduces a lock to avoid the crashes by serializing the affected accesses.

Calls to an instancer's Populate method may happen in parallel when there are multiple prototypes or prototype prims. When the instancer is itself in the prototype of another instancer, those parallel calls lead to parallel calls to _PopulateInstances on the parent instancer. These parallel calls to the parent instancer are always identical. instance modification is idempotent, but creation and deletion are not. Therefore, if the updates require the parent instancer to create or delete an instance of the child instancer, the parallel calls may lead to duplicated instances in the case of instance creation, or to a crash in the case of instance deletion, since prman is not tolerant of deleting an already-deleted instance.

To avoid the issue, which should be rare, we drop duplicative _PopulateInstances() calls from the same child instancer using a _LockingMap of mutexes, keyed by prototypePrimPath (i.e., the child instancer's path). Any calls received from the same child instancer during the duration of the lock are assumed to be redundant and are dropped. But parallel calls coming in from different child instancers will establish separate locks and therefore will not contend. With this new lock, contention discards unnecessary calls, so contention should not affect performance. Performance impact even in scenes with heavy nested instancing is expected to be minimal.

This change also switches all mutexes used by HdPrmanInstancer from std::mutex to tbb::spin_rw_mutex, which gives slightly better performance overall and allows us to avoid an exclusive lock in _AcquireGroupId() when the desired group id already exists.

(Internal change: 2305677)