Add user processing function and UsdUtilsLocalizeAsset
This change builds upon the newly added [UsdUtils_AssetLocalizationPackage](2301707).  This adds a new method `UsdUtilsLocalizeAsset` which will localize an asset to a directory on disk.  It also introduces the concept of a user processing function which allows users to provide a function which can modify asset paths and dependencies while leveraging built in asset discovery and traversal code.

The user processing function receives and returns a `UsdUtilsDependencyInfo` class which contains an asset path and vector of additional dependencies.

When returned as a parameter from a user processing function, the asset path value is checked by the localization system to see if it differs from the original authored value. If the returned value is set to and empty string, the asset will be removed from the layer and not included in the resulting package.  If this value differs from what what was originally authored into the layer, the path will be updated with this new value. Additionally, the newly specified asset will be included in the package and searched for additional dependencies if it can be opened as a layer.

When passed into the user processing function, if the dependencies array is populated, then the asset path resolved to one or more values, such as in the case of UDIM tiles or clip asset path template strings. When this structure is returned from a processing function, the paths contained within will be included in packaged output in addition to the value specified in the dependency's asset path. Any paths that can be opened as layers will be recursively searched for further dependencies.

(Internal change: 2305918)