pcp: Fix payload inclusion when computing ancestral opinions
Pcp was using the wrong path to determine if a payload
should be included when performing recursive prim indexing
to compute ancestral opinions for a requested prim index.
The correct "payload inclusion path" is the path of the
requested prim index, since that's the path of the prim
in the composed scenegraph that a user would be interacting
with. Pcp had been using the "rootSite" member of
Pcp_PrimIndexer, which in most cases contains that path,
but does not when computing ancestral opinions.

When evaluating payloads at a node, Pcp now computes the
payload inclusion path by translating the node's path
to the root of the requested prim index. This gets the
desired behavior when computing ancestral opinions.

This fix also replaces a somewhat complicated special
case in payload evaluation that forced ancestral payloads
to always be included when encountered in the context of
a subroot reference or payload, because in certain
situations there isn't a path in the composed scenegraph
that corresponds to this payload. The new path translation
logic detects this case more generally: if the path
translation fails, we must be in the situation described
above, so we know we must include the payload.

Lastly, this change also fixes an issue where the inclusion
of ancestral payloads for a prim index could be affected
by putting a path in the payload inclusion set that just
happened to match the path of the ancestral payload.

(Internal change: 2306329)