[hdSt, hgiVulkan] Pass bool indicating if draw batch is first draw batch in a graphics cmds submission to HdSt_DrawBatch::ExecuteDraw, allowing the pipeline draw batch and related drawing functionality know whether it should clear or not on a given draw command.
With Vulkan, the clearing of attachments will typically happen at the beginning
of a render pass. A Vulkan render pass is begun in HgiVulkanGraphicsCmds before
drawing. Each pipeline draw batch (which corresponds to one graphics pipeline)
will submit one draw command, but we can have multiple draw batches submit
draws to the same graphics cmds before submission of the graphics cmds. Thus,
we will begin multiple render passes per graphics cmds submission. However, if
the graphics cmds require a clearing of attachments, we really only want that
clearing to happen once, and not EVERY time we begin a render pass, otherwise
we will clear any drawing we just did.

In this change, we pass a bool indicating if the draw batch is the first in the
command buffer's draw batches (and thus the first in a graphics cmd submission)
to HdSt_DrawBatch::ExecuteDraw. HdSt_PipelineDrawBatch can then use this
information in hashing and creating the HgiGraphicsPipeline, which will then
correctly know if it should clear attachments or not. HgiVulkanGraphicsCmds can
then check with HgiVulkanGraphicsPipeline whether it should clear when it
begins a render pass.

(Internal change: 2306451)