Allow multiple USD in the same process on MacOS (#93)
This PR allows loading multiple usd libraries into the same process on macos (assuming they are uniquely namespaced and plugins are loaded from different locations etc).

It's implemented by driving the ArchConstructorEntry's version field through a preprocessor define so that the logic for calling them can distinguish entries from the two different builds and only call the appropriate ones.

This is arguably a hack abusing the field for something it should not be used for but if we change the layout of the struct it will
break cases where we don't have control of the other usd instance than the one we are building ourselves since the struct is
at play in both usd builds.

Adds the ability to load multiple usd libraries into the same process on macos.