usd: Fixes for UsdInherits::GetAllDirectInherits
This function relied on Pcp not culling certain inherit
nodes so that their associated site paths would be
included in the result, even if there were currently
no opinions at that path. However, this assumption
failed in (at least) cases that involved a mix of
specializes and inherit arcs. This was discovered by
running testUsdInherits with the PCP_CULLING env
setting set to 0.

We now use the expanded prim index to find the desired
inherit paths to ensure that all possible sources of
opinions are present.

This change also fixes a bug where paths in non-local
layer stacks would be included in the result.

(Internal change: 2307606)