Fix plugInfo reading when path contains regex characters
The previous code was failing to glob plugInfos when the path contained
special characters such as parentheses. For instance, given the path
/foo(bar)/**/plugInfo.json
the previous code was generating a regex
/foo(bar)/.*/plugInfo.json
which was not matching anything. This commit escapes the entire path,
including the non-wildcarded prefix. It also adds the missing special
characters in the escape function.