pcp: Per-node tracking of restricted opinion depth
Pcp now tracks the namespace depth at which a node was
restricted from contributing opinions to the composed
prim. This is needed by an upcoming fix to composition
that examines ancestral variant opinions from various
nodes in a prim index. The "contribution restriction
depth" will allow us to determine the namespace location
at which we can/must stop looking for variants to
compose into the prim index.

This change adds a new unit test (testPcpPrimIndex.py)
to exercise this functionality. This test uses helper
code that was refactored out of testPcpExpressionComposition.py
and into Pcp's own Python module as Pcp._TestPrimIndex.
This follows the same pattern as Pcp._TestChangeProcessor,
which is another test-only helper in the Python module.

(Internal change: 2308019)