pcp: Fix layer muting bug with file format targets
Pcp did not consider the file format target used for prim
index computations when determining whether a given layer
should be considered muted. So, if a user added a layer
identifier with an embedded file format target to the muted
layer set (e.g., "foo.usda:SDF_FORMAT_ARGS:target=x"), and
"foo.usda" was seen during a prim index computation where
the file format target was set to "x", this layer would
not be considered muted when it should have.

When computing the "canonical" identifier used to check
if a layer is muted, Pcp now strips out the embedded
file format target argument if it's present and matches
the file format target used during prim indexing. In
the example above, layer muting would work as expected
because the canonical identifier added to the muted
layer set would be just "foo.usda".

(Internal change: 2309311)