[hgiVulkan] Updated use of Vulkan SDK header implementations
Both SPIRV-Reflect and VMA (Vulkan Memory Allocator) are published
in the Vulkan SDK as header source that can be included by client
code.

We now include these implementations from the SDK at compile time
instead of maintaining static copies as part of the hgiVulkan source.

Updated VERSIONS.md to indicate the version of the Vulkan SDK
that we are building and testing against:
   vulkansdk-1.3.243.0

This is a simplified version of:
  https://github.com/PixarAnimationStudios/OpenUSD/pull/2563

Fixes #2563

(Internal change: 2309689)