UsdNamespaceEditor now finds all relationship targets and attribute connections for properties on the stage the stage that refer to the paths changed by a namespace edit and attempts to modify them to refer to the updated path (or remove the target altogether in the case of namespace deletion).
The test here currently only handles the basic case where all target and
connection paths exist in specs in the root layer stack. The
UsdNamespaceEditor does handle targets and connections that are defined
across composition arcs, but the test cases for this will be added in
an upcoming change.

(Internal change: 2310948)