pcp: Defer node culling passes to the end of indexing
When recursively computing prim indexes to incorporate
ancestral opinions due to composition arcs, Pcp would scan
and mark cullable nodes in those indexes before adding them
to the "final" requested prim index. These culling passes
assumed that all ancestral composition arcs had already been
processed.

An upcoming fix will defer the handling of ancestral variant
arcs, breaking this assumption. In anticipation of this, this
change defers the culling pass until the final prim index has
been fully computed. This has the side benefit of removing
some semi-complicated logic that would "uncull" nodes in
certain cases.

This change causes the TrickySpookyVariantSelectionInClass
test case to fail, but the behavior that test was intended to
verify was already broken to begin with; the test just happened
to be set up in such a way to mask that failure. That behavior
will be fixed by the upcoming change, so we just disable the
test for now.

Besides that, I don't expect any practically noticeable changes
in behavior or performance. Pcp did not rely on node culling
for any optimizations during indexing, and we should wind up
with the same nodes being culled and removed after indexing
is completed. Landing this change first will help verify that
and allow us to track unexpected regressions if they occur.

(Internal change: 2311004)