PcpMuseum test output/baseline cleanup: removing all remaining "_pcp" suffixes from output file and baselines. This required renaming existing baseline files using the "_pcp" suffix to text files without the suffix.
In some cases, the suffixless baseline didn't exist already so a simple
perforce rename was all that was required.

In other cases the suffixless file did exist but was not being used
anymore. These files were updated to match the "_pcp" baseline (if they
didn't already) and the "_pcp" baseline deleted.

(Internal change: 2311696)