[hdPrman] Use shutter curve on the camera when available.
- Update HdPrmanCamera to query and cache shutter curve attributes.
- Rework HdPrman_CameraContext to use the camera's shutter curve when available. It no longer needs to cache any of this state nor expose API to set it.
- Add unit test for various shutter interval and curve scenarios. See movingSphere.usda for notes on having motion blur show up correctly when using usdview.
- Update existing motion blur tests to use the fallback shutter curve for batch rendering so the baselines can remain unchanged.

Note that usdImaging wasn't updated to accomplish this. The reason this works is because the 1.0 and 2.0 getters (UsdImagingDelegate::GetCameraParamValue and UsdImagingDataSourceCameraPrim) fallback to querying the attribute on the USD prim. The shutter curve attributes are auto-applied by the UsdRi PxrCameraAPI schema on any UsdGeomCamera prim and thus always available when using hdPrman with USD.

It would however be useful to add an UsdImagingAPISchemaAdapter for PxrCameraAPI to allow its attributes to be inspectable in the Hydra Scene Browser.

(Internal change: 2313615)