pcp: Ensure implied opinions are considered for ancestral variants
This change fixes a long-standing composition bug where
variant selections authored in implied inherit locations
would be ignored during recursive prim indexing for
ancestral opinions. This was because ancestral variants
were evaluated immediately, before implied inherit arcs
that might have contributed opinions were added to the graph.

Pcp now defers the evaluation of ancestral variants as late
as possible to ensure that all possible sources of opinions are
available when searching for variant selections. Conceptually,
this is the same thing Pcp does for "direct" variant selections.
However, this is trickier to implement because it requires Pcp
to evaluate variants on a node at a different level of namespace
than the node's current site.

This approach means that variants will never be evaluated in
recursive prim indexing calls. This allows us to remove some
complicated code that was needed to deal with traversing a
prim index across those recursive calls and replace it with a
straightforward strength-order traversal of a single prim index.

The TrickySpookyVariantSelectionInClass museum case contains
examples of where this occurs. This test began passing several
years ago, leading us to believe the bug had been fixed by some
other change. It turns out that we were just getting lucky
(or unlucky) and the test was set up in a way that just
happened to get the right answer, but for the wrong reason.
This test had been disabled in a previous change, but with
this fix we can reenable it.

(Internal change: 2314080)