Extend HdStResourceRegistry to support a sub resource registry.
A sub resource registry is an additional HdResourceRegistry defined
externally that can manage resources whose definitions and
implementations are not known to HdStResourceRegistry, but whose
lifetime is tied to an HdStResourceRegistry.

Additionally HdStResourceRegistry will invoke Commit and GarbageCollect
on any sub resource registries it has ownership of whenever those
functions are invoked on it.

(Internal change: 2315307)